services/extractor_service/requirements.txt
requests
beautifulsoup4
confluent-kafka
psycopg2-binary
python-dateutil
selenium==4.12.0
